pyproject.toml
[project]
name = "trabalho-disciplina-eprompts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "guardrails-ai>=0.5.15",
    "openai>=1.56.2",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "unicode>=2.9",
    "unidecode>=1.3.8",
]
